requirements.txt
pyspark==3.5.6
delta-spark==3.2.1
requests==2.32.5
pandas>=2.0.0
numpy>=1.22.0
loguru>=0.7.0
mock>=5.0.0
pytest>=8.0.0
